% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backwardPhenotypeSelection.R
\name{backwardPhenotypeSelection}
\alias{backwardPhenotypeSelection}
\title{Full Annotation Using Shape-constrained Trees}
\usage{
backwardPhenotypeSelection(
  projectPath,
  startingPhenotype,
  inputModelStr,
  metaDataDF
)
}
\arguments{
\item{projectPath}{A path to a directory containing a completed FAUST analysis.
The projectPath directory is expected to have a faustData sub-directory, which
contains a completed faust analysis.}

\item{startingPhenotype}{A character vector containing the phenotype of interest.
This must match a column in the file 'faustCountMatrix.rds' that is produced
by a complete faust run.}

\item{inputModelStr}{A string containing a model formula. This formula is used
to guide the backward selection. The LHS of the formula must read
'cbind(childCount,(parentCount-childCount))'. The RHS of the formula
(after the '~' character) must match column names of the metaDataDF.
The variable of interest must occupy the first position of the}

\item{metaDataDF}{A data frame with all relevant meta data (except the counts)
needed to fit the model formula described in the 'inputModelStr' parameter.
One variable must be called 'sampleName', and the entries must match folders
contained in projectPath/faustData/sampleData.}
}
\value{
A list with two slots.

Slot 'selections': the markers chosen at each selection stage.

Slot 'pvalues': the associated pvalues.
}
\description{
backwardPhenotypeSelection
}
