% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotypeSelection.R
\name{getCountsForTargetMarkers}
\alias{getCountsForTargetMarkers}
\title{getCountsForTargetMarkers}
\usage{
getCountsForTargetMarkers(
  projectPath,
  referencePhenotype,
  targetMarkers,
  metaDataDF
)
}
\arguments{
\item{projectPath}{A path to a directory containing a completed FAUST analysis.
The projectPath directory is expected to have a faustData sub-directory, which
contains a completed faust analysis.}

\item{referencePhenotype}{A character vector specifying the phenotype of interest.
The expression of this phenotype is used by the target markers to derive the count
matrix. This phenotype must match a column in the file 'faustCountMatrix.rds'
that is produced by a complete faust run.}

\item{targetMarkers}{The markers of interest in the sub-phenotype.}

\item{metaDataDF}{A data frame with all relevant meta data (except the counts)
needed to fit the model formula described in the 'inputModelStr' parameter.
One variable must be called 'sampleName', and the entries must match folders
contained in projectPath/faustData/sampleData.}
}
\value{
A data frame with counts for the sub-phenotype consisting of the target
markers. All meta data in the metaDataDF is merged on.
}
\description{
getCountsForTargetMarkers
}
