% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMMIForTargetMarker.R
\name{getMMIForTargetMarker}
\alias{getMMIForTargetMarker}
\title{Full Annotation Using Shape-constrained Trees}
\usage{
getMMIForTargetMarker(
  targetMarker,
  projectPath = normalizePath("."),
  debugFlag = FALSE
)
}
\arguments{
\item{targetMarker}{The marker of interest. Function will construct a matrix
of median marker for this target marker, assuming it was an active channel in the
initial faust analyiss.}

\item{projectPath}{The filepath locating the faustData directory of the completed
faust analysis.}

\item{debugFlag}{Bolean value. Set to TRUE to print sample processing info.}
}
\value{
A sample-by-phenotype matrix of median marker intensities for the
target marker.
}
\description{
Once FAUST has been used to generate a count matrix, this function
can be used to create a sample-by-phenotype matrix of median marker intensities
for a target marker of interest. The target marker must be one of the active
channels in the initial analysis.
}
\examples{

#getMMIForTargetMarker(targetMarker="IFNg",projectPath="/home/user",debugFlag=TRUE)

}
