% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeAnnotationEmbedding.R
\name{makeAnnotationEmbedding}
\alias{makeAnnotationEmbedding}
\title{Full Annotation Using Shape-constrained Trees}
\usage{
makeAnnotationEmbedding(projectPath, sampleNameVec, vizType = "uniform")
}
\arguments{
\item{projectPath}{A path to a directory containing a completed FAUST analysis.
The projectPath directory is expected to have a faustData sub-directory, which
contains a completed faust analysis.}

\item{sampleNameVec}{A character vector listing samples to embed. Each entry
in the vector must match a directory name in projectPath/faustData/sampleData.}

\item{vizType}{A string listing the type of annotation embedding to perform.
The default value is "uniform", which causes the distance between any two
annotations is uniform across coordinates. The other supported value is
"depthscore", which scales annotation coordinates by their normalized depth score.}
}
\value{
A data frame with the following columns:

Column \code{umapX} and \code{umapY}: the x-y coordinates of the embedding.

Column \code{sampleOfOrigin}: character vector listing the sample of origin.

Column \code{faustLabels}: the faust annotation of the observation.

Column \verb{marker name}: the observed expression data, windsorized at the
1st and 99th percentile, and scaled to the unit interval.

Column \verb{marker name_annot}: the FAUST annotation for the marker.
}
\description{
makeAnnotationEmbedding produces an annotation embedding for samples analyzed
by FAUST.
}
\details{
The annotation embedding is constructed by first windsorizing each sample
to the 1st and 99th percentile. Then, the observed measurements for each FAUST annotation
are normalized to mean 0 unit SD (if >1). Finally the normalized annotation measurements
are translated by coordinate onto annotation landmarks.

After transforming each sample in this way, all listed samples are concatenated.
The concatendated array is embedded using the UMAP algorithm with the qualitative
parameter settings suggested in doi:10.1038/nbt.4314.
}
