% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkAFPlotsForExpUnit.R
\name{mkAFPlotsForExpUnit}
\alias{mkAFPlotsForExpUnit}
\title{Fast Annotation Using Shape-constrained Trees}
\usage{
mkAFPlotsForExpUnit(
  projectPath,
  expUnit,
  threadNum = 1,
  debugFlag = FALSE,
  maxRidgesAtDepth = Inf
)
}
\arguments{
\item{projectPath}{An absolute path to a directory on your system. Output from the
FAUST pipeline is written to the directory "projectPath/faustData".}

\item{expUnit}{The experimental unit you want to plot.}

\item{threadNum}{Number of threads to use.}

\item{debugFlag}{Boolean value. Set to TRUE to print status of plotting code.}

\item{maxRidgesAtDepth}{Numeric value. Maximum number of ridgelines to plot at any depth.
In very large annotation forest, this can be used to prevent overplotting in derived
ridgeline plot.}
}
\description{
mkAFPlotsForExpUnit produces joyplot of marginal and conditional densities along with their gate locations
for the specified experimental unit.
}
