% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotClusterRidgesInSample.R
\name{plotClusterRidgesInSample}
\alias{plotClusterRidgesInSample}
\title{Fast Annotation using Shape-constrained Trees}
\usage{
plotClusterRidgesInSample(
  clusterPhenotype,
  sampleName,
  outputDirName,
  projectPath,
  plottingDevice = "pdf"
)
}
\arguments{
\item{clusterPhenotype}{A text string corresponding to a cell population
determined by FAUST. Must be taken from the
projectPath/faustData/faustCountMatrix.rds column names.}

\item{sampleName}{A text string corresponding to the experimental
sample in which you want to plot clusterPhenotype ridgelines.
Must be a directory name in projectPath/faustData/sampleData.}

\item{outputDirName}{A text string which will be create a directory
to store plot output. The directory will be placed in
projectPath/faustData/plotData/clusterRidges.}

\item{projectPath}{An absolute path to the directory where you
ran the FAUST.}

\item{plottingDevice}{string with device for saving graphical output.
By default it is set to "pdf".}
}
\value{
plotClusterRidgesInSample returns a null value on completion.
The main output is a file of type plottingDevice located in
"projectPath/faustData/plotData/clusterRidges/outputDirName"
that shows ridgelines of the clusterPhenotype distribution set against the
entire distribution of the sampleName for each marker used to define
the phenotype.
}
\description{
This function plots the
strategy for a population found by the FAUST pipeline.
}
\examples{
\dontrun{
myProjectPath <- "/path/to/directory/housing/faustData"
countMatrix <- readRDS(paste0(myProjectPath,"/faustData/faustCountMatrix.rds"))
allSnames <- list.files(file.path(myProjectPath,
                                  "faustData",
                                  "sampleData"))

plotClusterRidgesInSample(
    clusterPhenotype=colnames(countMatrix)[1],
    sampleName=allSnames[1],
    outputDirName=testPlot,
    projectPath=myProjectPath,
    plottingDevice="pdf"
)
}
}
