% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFaustGates.R
\name{plotFaustGates}
\alias{plotFaustGates}
\title{Fast Annotation using Shape-constrained Trees}
\usage{
plotFaustGates(cellPop, sampleName, projectPath = ".")
}
\arguments{
\item{cellPop}{A text string corresponding to a cell population
determined by the FAUST pipeline. Must be taken from the
projectPath/faustData/faustCountMatrix.rds column names.}

\item{sampleName}{A text string corresponding to the experimental
sample in which you want to plot cellPop's gating strategy. Must be
taken from the projectPath/faustData/faustCountMatrix.rds row names.}

\item{projectPath}{An absolute path to the directory where you
ran the FAUST pipeline.}
}
\value{
plotFaustGates returns a null value on completion. The main output is a pdf
located in
"projectPath/faustData/plotData/gatingStrats/cellPops"
that shows the gating strategy.
}
\description{
This function plots the depth-score-order gating
strategy for a population found by the FAUST pipeline.
}
\examples{
\dontrun{
myProjectPath <- "/path/to/directory/housing/faustData"
countMatrix <- readRDS(paste0(myProjectPath,"/faustData/faustCountMatrix.rds"))
plotFaustGates(rownames(countMatrix)[1],colnames(countMatrix)[1],myProjectPath)
}
}
