% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotypeSelection.R
\name{backwardPhenotypeSelection}
\alias{backwardPhenotypeSelection}
\title{backwardPhenotypeSelection}
\usage{
backwardPhenotypeSelection(
  projectPath,
  startingPhenotype,
  selectionModel,
  metaDataDF,
  markersInParentPhenotype = NULL,
  fixedMarkers = NULL
)
}
\arguments{
\item{projectPath}{A path to a directory containing a completed FAUST analysis.
The projectPath directory is expected to have a faustData sub-directory, which
contains a completed faust analysis.}

\item{startingPhenotype}{A character vector specifying the phenotype of interest.
This must match a column in the file 'faustCountMatrix.rds' that is produced
by a complete faust run.}

\item{selectionModel}{A function that takes a single parameter "inputData"
and returns a single numeric value, the "pvalue". The "inputData" is a
data frame guaranteed to contain two columns: "childCount", the per-sample
counts of the phenotype implied by the current marker combination, and
"parentCount", the per-sample counts implied by the markersInParentPhenotype
parameter. All remaining values are the per-sample values of variables
supplied by the metaDataDF.}

\item{metaDataDF}{A data frame with all relevant meta data (except the counts)
needed to fit the model formula described in the 'selectionModel' parameter.
One variable must be called 'sampleName', and the entries must match folders
contained in projectPath/faustData/sampleData.}

\item{markersInParentPhenotype}{The subset of markers in the startingPhenotype
used to compute the sample-by-sample parentCount. NULL by default, leads
to using all observations in each sample as the parentCount. Example:
if the startingPhenotype contains "CD3+ CD4+ CD8-", setting this parameter to
c("CD3","CD4","CD8") will cause the parent count in the GLMM to be total
CD4 T cells rather than all cells in the sample.}

\item{fixedMarkers}{The subset of markers in the startingPhenotype that are
fixed during backwards selection: they are never candidates for marginalization
from the phenotypes. NULL by default, allowing all markers in the phenotypes
to be candidates for marginalization. Example: if the startingPhenotype contains
"CD3+ CD4+ CD8-", setting this parameter to c("CD3","CD4","CD8") will cause every
phenotype considered along the selection path to contain the sub-phenotype
"CD3+ CD4+ CD8-".}
}
\value{
A list with four slots.

Slot 'selections': the markers chosen at each selection stage.

Slot 'phenotypes': the phenotype associated with the selected markers.

Slot 'pvalues': the associated pvalues.

Slot \code{encodingData}: data used to construct phenotype queries based on the
results.
}
\description{
backwardPhenotypeSelection
}
